requirements.txt
aiogram==3.7.0
sqlalchemy==2.0.27
asyncpg==0.29.0
python-dotenv==1.0.0
apscheduler==3.10.4
aiohttp==3.9.1
qrcode==7.4.2
pillow==10.1.0